--- src/main/java/com/epam/rd/autotasks/springstatefulcalc/calculator/CalculatorFactory.java
@@ -1,8 +1,0 @@
-package com.epam.rd.autotasks.springstatefulcalc.calculator;
-
-import com.epam.rd.autotasks.springstatefulcalc.analyzer.Analyzer;
-import com.epam.rd.autotasks.springstatefulcalc.service.CalculatorService;
-
-public interface CalculatorFactory {
-    Calculator getCalculator(Analyzer analyzer, CalculatorService calculatorService);
-}
