+++ src/main/java/com/epam/rd/autotasks/springstatefulcalc/config/WebCalculatorConfig.java
@@ -3,12 +3,13 @@ package com.epam.rd.autotasks.springstatefulcalc.config;
 import com.epam.rd.autotasks.springstatefulcalc.analyzer.Analyzer;
 import com.epam.rd.autotasks.springstatefulcalc.analyzer.analyzerFactory.LexemeAnalyzerFactoryImpl;
 import com.epam.rd.autotasks.springstatefulcalc.calculator.Calculator;
-import com.epam.rd.autotasks.springstatefulcalc.calculator.calculatorFactory.WebCalculatorFactoryImpl;
+import com.epam.rd.autotasks.springstatefulcalc.calculator.webCalculator.WebCalculatorImpl;
 import com.epam.rd.autotasks.springstatefulcalc.service.CalculatorService;
-import com.epam.rd.autotasks.springstatefulcalc.service.WebCalculatorServiceImpl;
+import com.epam.rd.autotasks.springstatefulcalc.service.calculatorService.WebCalculatorServiceImpl;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Scope;
 
 @Configuration
 @ComponentScan("com.epam.rd.autotasks.springstatefulcalc")
@@ -25,7 +26,8 @@ public class WebCalculatorConfig {
     }
 
     @Bean
+    @Scope("prototype")
     public Calculator calculator() {
-        return new WebCalculatorFactoryImpl().getCalculator(lexemeAnalyzer(), calculatorService());
+        return new WebCalculatorImpl(lexemeAnalyzer(), calculatorService());
     }
 }
