+++ src/main/java/com/epam/rd/autotasks/springstatefulcalc/service/CalculatorService.java
@@ -4,4 +4,10 @@ import java.util.Map;
 
 public interface CalculatorService {
     String getExpressionWithParamValue(Map<String, String> cacheData);
+
+    boolean isDigit(String value);
+
+    boolean isParameterHasOverLimitValue(String paramValue);
+
+    boolean isGoodFormatExpression(String expression);
 }
